+++ src/extension.ts
@@ -9,4 +9,6 @@ export function activate(context: vscode.ExtensionContext) {
 export function deactivate(context: vscode.ExtensionContext) {
   context.workspaceState.update("displayed", undefined);
   context.workspaceState.update("hidden", undefined);
+  context.workspaceState.update("subDisplayed", undefined);
+  context.workspaceState.update("subHidden", undefined);
 }